👖⏱ Checkpoint ./src/client.jsx:8/1638 ./src/app.jsx:8/4174 ./views/index.ejs:8/2684
