+++ src/client.jsx
@@ -1,24 +1,16 @@
-/* globals EDITOR_URL */
+/* globals EDITOR_URL, ENVIRONMENT, PROJECT_DOMAIN */
 import './polyfills.js';
 import * as Sentry from '@sentry/browser';
 
 import React from 'react';
 import {render} from 'react-dom';
-import {BrowserRouter} from 'react-router-dom';
-
-import {CurrentUserProvider} from './presenters/current-user.jsx';
-import {UserPrefsProvider} from './presenters/includes/user-prefs.jsx';
-import {DevTogglesProvider} from './presenters/includes/dev-toggles.jsx';
-import {Notifications} from './presenters/notifications.jsx';
-
-import Router from './presenters/pages/router.jsx';
+import App from './app.jsx';
 
 // First things first -- let's bring our error collection online:
 try {
   Sentry.init({
     dsn: 'https://029cb06346934232bbc4ea4f4c16f1b7@sentry.io/1247156',
     environment: window.ENVIRONMENT,
-    serverName: window.PROJECT_DOMAIN,
     beforeSend(event) {
       if (event.user) {
         // Don't send user's email address
@@ -29,23 +21,14 @@ try {
   });
 
   Sentry.configureScope((scope) => {
-    scope.setUser({tags: {bootstrap: true}});
+    scope.setUser({tags: {
+      bootstrap: true,
+      PROJECT_DOMAIN,
+    }});
   });
 } catch (error) {
-
-const App = () => (
-  <BrowserRouter>
-    <Notifications>
-      <UserPrefsProvider>
-        <DevTogglesProvider>
-          <CurrentUserProvider>
-            {api => <Router api={api}/>}
-          </CurrentUserProvider>
-        </DevTogglesProvider>
-      </UserPrefsProvider>
-    </Notifications>
-  </BrowserRouter>
-);
+  console.warn("Error bringing Sentry online", error);
+}
 
 // Here's a bunch of browser support tests
 // If any of them don't work we can't run in this browser
@@ -59,9 +42,10 @@ const func = (f, ...args) => f(...args); // Can we define arrow functions?
 func(async arg => await arg, Promise.resolve()); // Can we do async/await?
 /* eslint-enable no-unused-vars */
 
-// Assuming none of them threw, set the global
-// This will get used to check for compatibility
-// If it isn't there the browser is unsupported
+
+// Assuming none of them threw, set the global bootstrap function.
+// This will get used to check for compatibility in index.ejs
+// If it isn't there, the browser is unsupported.
 window.bootstrap = () => {
   if (location.hash.startsWith("#!/")) {
     window.location.replace(EDITOR_URL + window.location.hash);
