+++ views/index.ejs
@@ -37,15 +37,17 @@
   </head>
   <body> 
     <script>
-      var APP_URL = '<%= APP_URL %>';
       var API_URL = '<%= API_URL %>';
-      var EDITOR_URL = '<%= EDITOR_URL %>';
-      var CDN_URL = '<%= CDN_URL %>';
-      var GITHUB_CLIENT_ID = '<%= GITHUB_CLIENT_ID %>';
-      var FACEBOOK_CLIENT_ID = '<%= FACEBOOK_CLIENT_ID %>';
-      var EXTERNAL_ROUTES = <%- EXTERNAL_ROUTES %>;
-      var ZINE_POSTS = <%- ZINE_POSTS %>;
+      var APP_URL = '<%= APP_URL %>';
       var BUILD_COMPLETE = <%- BUILD_COMPLETE %>;
+      var CDN_URL = '<%= CDN_URL %>';
+      var EDITOR_URL = '<%= EDITOR_URL %>';
+      var ENVIRONMENT = '<%= ENVIRONMENT %>';
+      var EXTERNAL_ROUTES = <%- EXTERNAL_ROUTES %>;
+      var FACEBOOK_CLIENT_ID = '<%= FACEBOOK_CLIENT_ID %>';
+      var GITHUB_CLIENT_ID = '<%= GITHUB_CLIENT_ID %>';
+      var PROJECT_DOMAIN = '<%= PROJECT_DOMAIN %>';
+      var ZINE_POSTS = <%- ZINE_POSTS %>;
     </script>
 
     <div id="fallback" style="display: none">
